Add support for serialized JSON input #75
