+++ README.md
@@ -168,7 +168,7 @@ The following conceptual topics exist in the `PSRule` module:
 
 PSRule uses the following schemas:
 
-- [PSRuleOptions](schemas/PSRule-options-0.2.0.schema.json) - Schema for PSRule YAML configuration file.
+- [PSRuleOptions](schemas/PSRule-options-0.3.0.schema.json) - Schema for PSRule YAML configuration file.
 
 ## Changes and versioning
