+++ schemas/PSRule-options-0.3.0.schema.json
@@ -1,0 +1,179 @@
+{
+    "title": "PSRule options",
+    "$schema": "http://json-schema.org/draft-04/schema#",
+    "oneOf": [
+        {
+            "$ref": "#/definitions/options-v0.3.0"
+        }
+    ],
+    "definitions": {
+        "baseline": {
+            "type": "object",
+            "description": "Options that include/ exclude and configure rules.",
+            "properties": {
+                "ruleName": {
+                    "type": "array",
+                    "description": "Optionally filter rules by name.",
+                    "items": {
+                        "type": "string"
+                    },
+                    "uniqueItems": true
+                },
+                "exclude": {
+                    "type": "array",
+                    "description": "Specifies rules to exclude by name.",
+                    "items": {
+                        "type": "string"
+                    },
+                    "uniqueItems": true
+                },
+                "configuration": {
+                    "type": "object",
+                    "description": "A set of key/ value configuration options for rules."
+                }
+            },
+            "additionalProperties": false
+        },
+        "binding": {
+            "type": "object",
+            "description": "Configure property/ object binding options.",
+            "properties": {
+                "targetName": {
+                    "type": "array",
+                    "description": "Specifies one or more property names to bind TargetName to.",
+                    "items": {
+                        "type": "string"
+                    }
+                }
+            },
+            "additionalProperties": false
+        },
+        "execution": {
+            "type": "object",
+            "description": "Options that affect rule execution.",
+            "properties": {
+                "languageMode": {
+                    "type": "string",
+                    "description": "The PowerShell language mode to use for rule execution.",
+                    "enum": [
+                        "FullLanguage",
+                        "ConstrainedLanguage"
+                    ],
+                    "default": "FullLanguage"
+                },
+                "inconclusiveWarning": {
+                    "type": "boolean",
+                    "description": "Enable or disable warnings for inconclusive rules.",
+                    "default": true
+                },
+                "notProcessedWarning": {
+                    "type": "boolean",
+                    "description": "Enable or disable warnings for objects that are not processed by any rule.",
+                    "default": true
+                }
+            },
+            "additionalProperties": false
+        },
+        "input": {
+            "type": "object",
+            "properties": {
+                "format": {
+                    "type": "string",
+                    "description": "The input string format.",
+                    "enum": [
+                        "None",
+                        "Yaml",
+                        "Json"
+                    ],
+                    "default": "None"
+                }
+            }
+        },
+        "suppression": {
+            "type": "object",
+            "description": "Specifies suppression rules."
+        },
+        "options-v0.2.0": {
+            "properties": {
+                "baseline": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/baseline"
+                        }
+                    ]
+                },
+                "binding": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/binding"
+                        }
+                    ]
+                },
+                "execution": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/execution"
+                        }
+                    ]
+                },
+                "suppression": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/suppression"
+                        }
+                    ]
+                }
+            },
+            "additionalProperties": false
+        },
+        "options-v0.3.0": {
+            "properties": {
+                "baseline": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/baseline"
+                        }
+                    ]
+                },
+                "binding": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/binding"
+                        }
+                    ]
+                },
+                "execution": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/execution"
+                        }
+                    ]
+                },
+                "input": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/input"
+                        }
+                    ]
+                },
+                "suppression": {
+                    "type": "object",
+                    "oneOf": [
+                        {
+                            "$ref": "#/definitions/suppression"
+                        }
+                    ]
+                }
+            },
+            "additionalProperties": false
+        }
+    }
+}
