--- src/PSRule/Pipeline/PSObjectJsonConverter.cs
@@ -1,87 +1,0 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Management.Automation;
-
-namespace PSRule.Pipeline
-{
-    /// <summary>
-    /// Custom serializer to correctly convert PSObject properties to JSON instead of CLIXML.
-    /// </summary>
-    internal sealed class PSObjectJsonConverter : JsonConverter
-    {
-        public override bool CanConvert(Type objectType)
-        {
-            return objectType == typeof(PSObject);
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            if (!(value is PSObject obj))
-            {
-                throw new ArgumentException();
-            }
-
-            writer.WriteStartObject();
-
-            foreach (var property in obj.Properties)
-            {
-                // Ignore properties that are not readable or can cause race condition
-                if (!property.IsGettable || property.Value is PSDriveInfo || property.Value is ProviderInfo)
-                {
-                    continue;
-                }
-
-                writer.WritePropertyName(property.Name);
-                serializer.Serialize(writer, property.Value);
-            }
-
-            writer.WriteEndObject();
-        }
-
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            // Create target object based on JObject
-            var result = existingValue as PSObject ?? new PSObject();
-
-            // Read tokens
-            ReadObject(value: result, reader: reader);
-
-            return result;
-        }
-
-        private void ReadObject(PSObject value, JsonReader reader)
-        {
-            if (reader.TokenType != JsonToken.StartObject)
-            {
-                throw new Exception("Read json failed");
-            }
-
-            reader.Read();
-
-            string name = null;
-
-            // Read each token
-            while (reader.TokenType != JsonToken.EndObject)
-            {
-                switch (reader.TokenType)
-                {
-                    case JsonToken.PropertyName:
-                        name = reader.Value.ToString();
-                        break;
-
-                    case JsonToken.StartObject:
-                        var child = new PSObject();
-                        ReadObject(value: child, reader: reader);
-                        value.Properties.Add(new PSNoteProperty(name: name, value: child));
-                        break;
-
-                    default:
-                        value.Properties.Add(new PSNoteProperty(name: name, value: reader.Value));
-                        break;
-                }
-
-                reader.Read();
-            }
-        }
-    }
-}
