+++ src/PSRule/Pipeline/PipelineReciever.cs
@@ -1,12 +1,11 @@
-﻿using PSRule.Configuration;
-using System.Collections;
+﻿using Newtonsoft.Json;
+using PSRule.Configuration;
 using System.Collections.Generic;
 using System.IO;
 using System.Management.Automation;
 using YamlDotNet.Core;
 using YamlDotNet.Core.Events;
 using YamlDotNet.Serialization;
-using YamlDotNet.Serialization.NamingConventions;
 
 namespace PSRule.Pipeline
 {
@@ -20,20 +19,34 @@ namespace PSRule.Pipeline
             yield return targetObject;
         }
 
-        public static IEnumerable<PSObject> ConvertFromYaml(PSObject targetObject)
+        public static IEnumerable<PSObject> ConvertFromJson(PSObject sourceObject)
         {
-            if (!(targetObject.BaseObject is string))
+            if (!(sourceObject.BaseObject is string))
             {
-                return new PSObject[] { targetObject };
+                return new PSObject[] { sourceObject };
+            }
+
+            var result = new List<PSObject>();
+
+            var value = JsonConvert.DeserializeObject<PSObject[]>(sourceObject.BaseObject.ToString(), new PSObjectArrayJsonConverter());
+
+            return value;
+        }
+
+        public static IEnumerable<PSObject> ConvertFromYaml(PSObject sourceObject)
+        {
+            if (!(sourceObject.BaseObject is string))
+            {
+                return new PSObject[] { sourceObject };
             }
 
             var d = new DeserializerBuilder()
                 .IgnoreUnmatchedProperties()
-                .WithNamingConvention(new CamelCaseNamingConvention())
-                .WithNodeTypeResolver(new HashtableTypeResolver())
+                .WithTypeConverter(new PSObjectYamlTypeConverter())
+                .WithNodeTypeResolver(new PSObjectYamlTypeResolver())
                 .Build();
 
-            var reader = new StringReader(targetObject.BaseObject.ToString());
+            var reader = new StringReader(sourceObject.BaseObject.ToString());
             var parser = new Parser(reader);
 
             parser.Expect<StreamStart>();
@@ -42,8 +55,8 @@ namespace PSRule.Pipeline
 
             while (parser.Accept<DocumentStart>())
             {
-                var item = d.Deserialize<Hashtable>(parser: parser);
-                result.Add(new PSObject(item));
+                var item = d.Deserialize<PSObject>(parser: parser);
+                result.Add(item);
             }
 
             return result.ToArray();
