+++ tests/PSRule.Tests/PSRule.Tests.csproj
@@ -28,6 +28,9 @@
     <None Update="ObjectFromFile.yaml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+    <None Update="ObjectFromFile.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
